+++ CHANGELOG.md
@@ -4,8 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0]
+### Added
+- Make it possible to use a custom rates file to import
+
+### Fixed
+- Implemented Unit Test for the import script
+
+## [1.0.0]
 ### Added
 - Commandline script to import tax rates and create tax rules
 
-[Unreleased]: https://github.com/elgentos/magento2-eu-tax-rates
+[Unreleased]: https://github.com/elgentos/magento2-eu-tax-rates/tree/main
+[1.1.0]: https://github.com/elgentos/magento2-eu-tax-rates/compare/1.0.0...1.1.0
+[1.0.0]: https://github.com/elgentos/magento2-eu-tax-rates/releases/tag/1.0.0